Network katmanı düzenlendi. Çek-senet sayfaları geliştiriliyor.

+++ lib/view/main_page/alt_sayfalar/finans/cek_senet/cek_senet_tahsilati/model/save_cek_senet_model.dart
@@ -48,7 +48,7 @@ class  CekSenetkalemlerModel with _$CekSenetkalemlerModel {
     int? sira,
     String? tag,
     DateTime? tarih,
-    int? tutar,
+    double? tutar,
     DateTime? vadeTarihi,
   }) = _CekSenetkalemlerModel;
